Speed up native_safe_apic_wait_icr_idle()
Waiting for 100us before rechecking the ICR strikes me as slow.  What's
more, we have observed a 24-cpu system receive roughly 800 NMIs per cpu
per second.  Assuming each of them causes the sender to spin inside this
loop exactly once, the total time spent spinning is somewhere near 2s.
In other words, it looks as if we burn two cpus spinning in this loop.

Reduce the waiting time to 100ns, or roughly the time to go to DRAM.
Including rounding errors, that might cause the timeout to be slightly
increased - but I believe the timeout is a) pulled out of thin air and
b) only strikes if the system is pretty hosed anyway.

Signed-off-by: Joern Engel <joern@logfs.org>