iTCO: add pretimeout
The watchdog user can set a timer to fire before the actual watchdog
hits.  If it fires, a kernel messages is printed, the process holding
/dev/watchdog open gets SIGILL'ed and we do a sync - all in the hope
that as much debug output as possible can get salvaged before the actual
watchdog bites.

There is no hard guarantee that a pretimeout will happen before the
watchdog.  It can be usespace's fault that the watchdog fires, but it
may as well be a wedged kernel, which no longer has working timers.

Signed-off-by: Joern Engel <joern@logfs.org>