qla2xxx: Make recv_time and ctio_time monotonic, and move into struct se_cmd
Move these timestamps up into se_cmd so that they're accessible to LIO.

In upstream LIO there is no guarantee that they completion and request
are on the same cpu, so local_clock() isn't appropriate. Although we have
local patches which means that completions are sent out on the same cpu,
other (large) parts of the IO stack will run on different cpus, and rather
than convert between clocks we should just use the same one. So switch to
ktime_get_ts() instead.