BUG_ON(page_mapped(page));
On Fri, 20 Apr 2012, Hugh Dickins wrote:
> On Fri, 20 Apr 2012, Roland Dreier wrote:
> 
> > I do see where truncate_inode_page() calls
> > unmap_mapping_range() but is it possible we still have another
> > mapping surviving from the other process?  (I'm not sure that even
> > makes sense at ask)
> 
> Yes, given the BUG, it's quite likely that a mapping (pte) has
> survived in one or the other process; whereas unmap_mapping_range's
> job is to find them all and remove them.

I was wrong to go on about unmap_mapping_range() and the prio_tree:
bugs there should be irrelevant in this case.  They're important when
carrying out an actual truncation of an open file, but your backtrace
clearly showed the inode being evicted when it's unlinked - which
implies that nobody else had it still open or mapped by that time
(but perhaps there's a file or inode refcounting error somewhere).

> 
> I wonder if the mm of old process #1 or old process #2 is occasionally
> still being torn down at the time that you start up the new process #1.
> Maybe it would involve an additional process, looking at stats via /proc,
> holding on to the mm after the tasks themselves have gone.
> 
> Now, of course that shouldn't generate any BUG; but it does feel like
> a worthwhile line of enquiry, a good arena to search for bugs in.
> I'll sleep on it tonight!

No other insights came overnight, I'm afraid.

You suggested in your first mail that we could add more debug output.
Here's a patch against 2.6.39.4 to replace that BUG_ON by "Bad page"
info and continue running: because, crashing at that point, we have
no idea whether this is something occasionally afflicting a single
page of the 12GB (corruption of the radix_tree?  reuse of a page?),
or whether when it happens it actually hits every page of the 12GB
(as would be more likely if the file or inode refcounting is off).

This shouldn't spam your logs too much: the first 130 occurrences,
then just once per 1GB after that if every page is wrong.  There
may or may not be "Bad page state" reports interspersed, when and
if the pages are freed: those show 60 reports then wait a minute
(but for these radix_tree ones, I'd like to see at least two nodes).

Hugh