mpt2sas-reorder-smid-list-v1
In a high-IOPS workload, mpt2sas_base_get_smid_scsiio shows up in the top
20 cache misses.  This is because the data structure used for allocating
SMIDs is cache unfriendly; freed SMIDs are placed on the tail of the list,
guaranteeing a cache miss by the time we allocate it again.

By placing the freed SMID at the head of the list, we increase the
likelihood of it being cache-hot when it's used again.

See: <http://www.spinics.net/lists/linux-scsi/msg51610.html>